Implemented features for creating, retrieving, updating and  deleting part categories

+++ backend/src/cloud/cloudOperations.ts
@@ -1,33 +1,46 @@
 import { Bucket } from "@google-cloud/storage";
-import { UploadOptions } from "../Interface/interface";
+import { ImageResize, UploadOptions } from "../Interface/interface";
 import { errorHandler } from "../Utils/errorHandler";
 import { Response } from "express";
 import fs from "fs";
 import axios from "axios";
+import resizeImage from "../imageOptmization/resize";
+
+interface ImageResizeOptions {
+	height : number, 
+	width : number
+}
 
 async function uploadFileToCloud(
 	res: Response,
 	file: Express.Multer.File,
 	bucket: Bucket,
-	uploadOptions: UploadOptions
+	uploadOptions: UploadOptions,
+	resizeOptions ?: ImageResizeOptions
 ) {
-	return new Promise<string>((resolve, reject) => {
-		bucket.upload(file.path, uploadOptions, async (err, uploadedFile) => {
+	return new Promise<string>(async (resolve, reject) => {
+		let filePath = file.path;
+		if(resizeOptions){
+			const folderpath = file.path.split('\\')
+			const resizedPath = await resizeImage({imagePath : file.path, height : resizeOptions.height, width : resizeOptions.width, resizedImagePath : `${folderpath.slice(0, folderpath.length - 1)}`})
+			filePath = resizedPath
+			console.log('path' , filePath)
+		}
+		bucket.upload(filePath, uploadOptions, async (err, uploadedFile) => {
 			if (err) {
-				errorHandler(res, 500, "Error uploading file", {
-					errMessage: err.message,
-				});
 				reject(err);
 			}
 			await uploadedFile?.makePublic();
 			const fileUrl = uploadedFile?.publicUrl();
+			console.log('path inside : ', filePath)
+			console.log(fileUrl)
 
 			if (!fileUrl) {
-				errorHandler(res, 500, "Error while retrieving fileUrl");
 				reject(new Error("Error while retrieving fileUrl"));
 				return;
 			}
-			fs.unlinkSync(file.path); // removing the file from upload directory
+			resizeOptions && fs.unlinkSync(file.path)
+			fs.unlinkSync(filePath); // removing the file from upload directory
 			resolve(fileUrl);
 		});
 	});
@@ -72,8 +85,8 @@ async function deleteCloudFile(filePath: string, bucket: Bucket) {
 	});
 }
 
-async function updateCloudFile(res: Response, file: Express.Multer.File,filePathInStorage : string, bucket: Bucket, uploadOptions: UploadOptions) {
-	const fileSrc = await uploadFileToCloud(res, file, bucket, uploadOptions);
+async function updateCloudFile(res: Response, file: Express.Multer.File,filePathInStorage : string, bucket: Bucket, uploadOptions: UploadOptions, resizeOptions ?: {height : number, width : number}) {
+	const fileSrc = await uploadFileToCloud(res, file, bucket, uploadOptions,resizeOptions);
 	const isDeleted = await deleteCloudFile(filePathInStorage, bucket);
 	return {
 		fileSrc,
